+++ src/pages/Main.tsx
@@ -1,31 +1,18 @@
-import { useEffect, useState } from 'react';
-
 import Header from '../components/Header';
 import InputTodo from '../components/InputTodo';
 import TodoList from '../components/TodoList';
-import { getTodoList } from '../api/todo';
 
-export interface ITodoListData {
-  id: number;
-  title: string;
-}
+import useTodo from '../hooks/useTodo';
 
 const Main = () => {
-  const [todoListData, setTodoListData] = useState<ITodoListData[]>([]);
-
-  useEffect(() => {
-    (async () => {
-      const { data } = await getTodoList();
-      setTodoListData(data || []);
-    })();
-  }, []);
+  useTodo();
 
   return (
     <div className="container">
       <div className="inner">
         <Header />
-        <InputTodo setTodos={setTodoListData} />
-        <TodoList todos={todoListData} setTodos={setTodoListData} />
+        <InputTodo />
+        <TodoList />
       </div>
     </div>
   );
